Correct 'join' button press bug.
The join action was bound to the 'submit' event. The join button was an
anchor, not a button so when you pressed it it don't fire the form
event. It is now a button.

I also cleaned but the code. There is a var 'joinForm' which is a
jQuery object. This is to match 'chatForm' which was already there.