WIP
Removed the current spinner, will add a small one to the login screen itself.
The login screen acts more like a yahoo messenger login screen, you can't get past it if the nick is the same. If a password will be requested, probably a password box will be added to the username edit box.
Fixed a bug for how the client connects, made it asynchronous.
Cleaned up the code.